+++ generator.ts
@@ -8,13 +8,14 @@ import * as os from 'os';
 import * as mustache from 'mustache'
 import * as drillInRpt from './reports/drill-in'
 import {Crawler} from './crawler';
-import {DistinctSet} from './util';
+import * as lib from './project-reports-lib'
+import moment = require('moment');
 
 let sanitize = require('sanitize-filename');
 let clone = require('clone');
 
-import { CrawlingConfig, GeneratorConfiguration, ProjectIssue, ReportSnapshot, ReportConfig, ProjectData, ProjectReportBuilder, ReportDetails, IssueSummary, CrawlingTarget } from './interfaces'
-//import { url } from 'inspector';
+import {IssueList, ProjectIssue} from './project-reports-lib';
+import { CrawlingConfig, CrawlingTarget, GeneratorConfiguration, ReportSnapshot, ReportConfig, ProjectReportBuilder, ReportDetails } from './interfaces'
 
 export async function generate(token: string, configYaml: string): Promise<ReportSnapshot> {
     const workspacePath = process.env["GITHUB_WORKSPACE"];
@@ -51,7 +52,7 @@ export async function generate(token: string, configYaml: string): Promise<Repor
         report.timezoneOffset = report.timezoneOffset || -8;
 
         report.details = <ReportDetails>{
-            time: util.getTimeForOffset(snapshot.datetime, report.timezoneOffset)
+            time: moment().utcOffset(report.timezoneOffset).format("dddd, MMMM Do YYYY, h:mm:ss a")
         }
         report.details.rootPath = path.join(snapshot.rootPath, sanitize(report.name));
         report.details.fullPath = path.join(report.details.rootPath, snapshot.datetimeString);
@@ -150,7 +151,7 @@ export async function generate(token: string, configYaml: string): Promise<Repor
             // ----------------------------------------------------------------------
             let targetNames = reportSection.targets || report.targets;
 
-            let set = new DistinctSet(issue => issue.html_url);
+            let set = new IssueList(issue => issue.html_url);
             
             let targets: CrawlingTarget[] = [];
             for (let targetName of targetNames) {
@@ -164,7 +165,7 @@ export async function generate(token: string, configYaml: string): Promise<Repor
                     throw new Error(`Report target mismatch.  Target is of type ${target.type} but report section is ${reportGenerator.reportType}`);
                 }
 
-                let data: IssueSummary[] = await crawler.crawl(target);
+                let data: ProjectIssue[] = await crawler.crawl(target);
                 console.log(`Adding ${data.length} issues to set ...`);
                 set.add(data);
             }
