+++ interfaces.ts
@@ -1,4 +1,4 @@
-import { DistinctSet } from "./util";
+import { IssueList, ProjectIssue } from "./project-reports-lib";
 
 export interface GeneratorConfiguration {
     name: string,
@@ -12,7 +12,9 @@ export interface GeneratorConfiguration {
 export interface CrawlingTarget {
     type: 'project' | 'repo',
     htmlUrl: string,
-    columnMap: { [key: string]: string[] },    
+    columnMap: { [key: string]: string[] },
+    // only needed for type project
+    projectId?: number,  
 }
 
 export type CrawlingConfig = {[name: string]: CrawlingTarget};
@@ -61,104 +63,6 @@ export interface ProjectData {
     // stages: { [key: string]: ProjectIssue[] }
 }
 
-export interface IssueLabel {
-    name: string
-}
-
-export interface IssueCardEventProject {
-    project_id: number,
-    column_name: string,
-    previous_column_name: string,
-    stage_name: string,
-    previous_stage_name: string
-}
-
-export interface IssueEvent {
-    created_at: Date,
-    event: string,
-    assignee: IssueUser,
-    label: IssueLabel,
-    project_card: IssueCardEventProject,
-    //data: any
-}
-
-export interface IssueUser {
-    login: string,
-    id: number,
-    avatar_url: string,
-    url: string,
-    html_url: string
-}
-
-export interface IssueMilestone {
-    title: string,
-    description: string,
-    due_on: Date
-}
-
-export interface IssueComment {
-    body: string,
-    user: IssueUser
-    created_at: Date,
-    updated_at: Date
-}
-
-//
-// shallow issue for bug slicing and dicing
-//
-export interface IssueSummary {
-    title: string,
-    number: number;
-    html_url: string,
-    state: string,
-    labels: IssueLabel[],
-    assignee: IssueUser,
-    assignees: IssueUser[],
-    user: IssueUser,
-    milestone: IssueMilestone,
-    closed_at: Date,
-    created_at: Date,
-    updated_at: Date
-}
-
-export interface ProjectIssue extends IssueSummary {
-    labels: IssueLabel[],
-    assignee: IssueUser,
-    assignees: IssueUser[],
-    user: IssueUser,
-    milestone: IssueMilestone,
-    closed_at: Date,
-    created_at: Date,
-    updated_at: Date,
-
-    comments: IssueComment[],  
-    
-    //
-    // project stage fields we decorate on issues
-    //
-
-    // first added to the board on any column (no "from" column)
-    project_added_at: Date,
-    
-    // last occurence of moving to these columns from a lesser or no column
-    // example. if moved to accepted from proposed (or less), 
-    //      then in-progress (greater) and then back to accepted, first wins
-    project_proposed_at: Date,        
-    project_accepted_at: Date,
-    project_in_progress_at: Date,
-
-    // cleared if not currently blocked
-    project_blocked_at: Date,
-
-    // cleared if it moves out of done.  e.g. current state has to be done for this to be set
-    project_done_at: Date,
-
-    // current stage of this card on the board
-    project_stage: string,
-    
-    events: IssueEvent[]
-}
-
 export interface IssueParameters {
     state: string,
     milestone: string,
@@ -169,7 +73,7 @@ export interface ProjectReportBuilder {
     // a report accepts project data (and it's stages) or a list of issues from a repo (and it's stages)
     reportType: "project" | "repo" | "any";
     getDefaultConfiguration(): any;
-    process(config: any, data: ProjectData | DistinctSet, drillIn: (identifier: string, title: string, cards: ProjectIssue[]) => void): any;
+    process(config: any, data: IssueList, drillIn: (identifier: string, title: string, cards: ProjectIssue[]) => void): any;
     renderMarkdown(targets: CrawlingTarget[], processedData?: any): string;
     renderHtml(targets: CrawlingTarget[], processedData?: any): string;
 }
