+++ package.json
@@ -17,6 +17,7 @@
     "clone": "^2.1.2",
     "https-proxy-agent": "^5.0.0",
     "js-yaml": "^3.14.0",
+    "moment": "^2.27.0",
     "mustache": "^4.0.1",
     "sanitize-filename": "^1.6.3",
     "tablemark": "^2.0.0"
