+++ project-reports-lib.ts
@@ -1,5 +1,7 @@
-import {ProjectIssue, IssueSummary} from './interfaces';
 import * as url from 'url' 
+import * as clone from 'clone';
+import moment = require('moment');
+import { IssueLabelBreakdown } from './reports/repo-issues';
 
 // TODO: separate npm module.  for now it's a file till we flush out
 
@@ -22,13 +24,8 @@ export function repoPropsFromUrl(htmlUrl: string): RepoProps {
 //
 // filter cards by label
 //
-export function filterByLabel(issues: IssueSummary[], name: string): IssueSummary[] {
+export function filterByLabel(issues: ProjectIssue[], name: string): ProjectIssue[] {
     return issues.filter((card) => card.labels.findIndex(label => label.name.trim().toLowerCase() === name.toLowerCase()) >= 0);
-}
-
-export function diffDays(start: Date, end: Date): number {
-    const difference = end.getTime() - start.getTime();
-    return difference / (1000 * 60 * 60 * 24);
 }
 
 //
@@ -108,4 +105,306 @@ export function getProjectStageIssues(issues: ProjectIssue[]) {
     }
 
     return projIssues;
+}
+
+export interface IssueLabel {
+    name: string
+}
+
+export interface IssueCardEventProject {
+    project_id: number,
+    column_name: string,
+    previous_column_name: string,
+    stage_name: string,
+    previous_stage_name: string
+}
+
+export interface IssueEvent {
+    created_at: Date,
+    event: string,
+    assignee: IssueUser,
+    label: IssueLabel,
+    project_card: IssueCardEventProject,
+    //data: any
+}
+
+export interface IssueUser {
+    login: string,
+    id: number,
+    avatar_url: string,
+    url: string,
+    html_url: string
+}
+
+export interface IssueMilestone {
+    title: string,
+    description: string,
+    due_on: Date
+}
+
+export interface IssueComment {
+    body: string,
+    user: IssueUser
+    created_at: Date,
+    updated_at: Date
+}
+
+
+export interface ProjectIssue {
+    title: string,
+    number: number;
+    html_url: string,
+    state: string,
+    labels: IssueLabel[],
+    assignee: IssueUser,
+    assignees: IssueUser[],
+    user: IssueUser,
+    milestone: IssueMilestone,
+    closed_at: Date,
+    created_at: Date,
+    updated_at: Date
+
+    comments: IssueComment[],  
+    events: IssueEvent[],
+    
+    //
+    // project stage fields we decorate on issues
+    //
+
+    // first added to the board on any column (no "from" column)
+    project_added_at: Date,
+    
+    // last occurence of moving to these columns from a lesser or no column
+    // example. if moved to accepted from proposed (or less), 
+    //      then in-progress (greater) and then back to accepted, first wins
+    project_proposed_at: Date,        
+    project_accepted_at: Date,
+    project_in_progress_at: Date,
+
+    // cleared if not currently blocked
+    project_blocked_at: Date,
+
+    // cleared if it moves out of done.  e.g. current state has to be done for this to be set
+    project_done_at: Date,
+
+    // current stage of this card on the board
+    project_stage: string,
+}
+
+let stageLevel = {
+    "None": 0,
+    "Proposed": 1,
+    "Accepted": 2,
+    "In-Progress": 3,
+    "Done": 4
+}
+
+export class IssueList {
+    seen;
+    identifer;
+    items: ProjectIssue[];
+
+    // keep in order indexed by level above
+    // TODO: unify both to avoid out of sync problems
+    stageAtNames = [
+        'none',
+        'project_proposed_at',
+        'project_accepted_at',
+        'project_in_progress_at',
+        'project_done_at'
+    ]
+
+    constructor(identifier: (item) => any) {
+        this.seen = new Map();
+        this.identifer = identifier;
+        this.items = [];
+    }
+
+    // returns whether any were added
+    public add(data: any | any[]): boolean {
+        let added = false;
+        if (Array.isArray(data)) {
+            for (let item of data) {
+                let res = this.add_item(item);
+                if (!added) { added = res; }
+            }
+        }
+        else {
+            return this.add_item(data);
+        }
+
+        return added;
+    }
+
+    private add_item(item: any): boolean {
+        let id = this.identifer(item);
+
+        if (!this.seen.has(id)) {
+            this.items.push(item);
+            this.seen.set(id, item);
+            return true;
+        }
+
+        return false;
+    }
+
+    public getItem(identifier: any): ProjectIssue {
+        return this.seen[this.identifer];
+    }
+
+    public getItems(): ProjectIssue[] {
+        // call process
+        for (let item of this.items) {
+            this.processStages(item);    
+        }
+        
+        return this.items;
+    }
+
+    //
+    // Gets an issue from a number of days, hours ago.
+    // Clones the issue and Replays events (labels, column moves, milestones)
+    // and reprocesses the stages.
+    // If the issue doesn't exist in the list, returns null
+    //
+    public getItemAgo(identifier: any, amount: number, unit: 'days' | 'hours' ): ProjectIssue {
+        let issue = this.getItem(this.identifer);
+
+        if (!issue) {
+            return issue;
+        }
+
+        let momentAgo = moment().subtract(amount, unit);        
+
+        // clear everything we're going to re-apply
+        issue.labels = [];
+        delete issue.project_added_at;
+        delete issue.project_proposed_at;
+        delete issue.project_in_progress_at;
+        delete issue.project_accepted_at;
+        delete issue.project_done_at;
+        delete issue.project_stage;
+        
+        // stages and labels
+        let filteredEvents: IssueEvent[] = [];
+        let labelMap: { [name: string]: IssueLabel };
+
+        if (issue.events) {
+            for (let event of issue.events) {
+                if (momentAgo.isBefore(event.created_at)) {
+                    continue;
+                }
+
+                filteredEvents.push(event);
+
+                if (event.event === 'labeled') {
+                    labelMap[event.label.name] = event.label;
+                }
+                else if (event.event === 'unlabeled') {
+                    delete labelMap[event.label.name];
+                }
+            }
+        }
+        issue.events = filteredEvents;
+        for (let labelName in labelMap) {
+            issue.labels.push(labelMap[labelName]);
+        }
+        
+        this.processStages(issue);
+
+        // comments
+        let filteredComments: IssueComment[] = [];
+        for (let comment of issue.comments) {
+            if (momentAgo.isBefore(comment.created_at)) {
+                filteredComments.push(comment);
+            }
+        }
+        issue.comments = filteredComments;
+
+        return issue;
+    }
+
+    //
+    // Process the events to set project specific fields like project_done_at, project_in_progress_at, etc
+    // Call initially and then call again if events are filtered (get issue asof)
+    //
+    private processStages(issue: ProjectIssue): void {
+        // card events should be in order chronologically
+        let currentStage: string;
+        let doneTime: Date;
+        let addedTime: Date;
+
+        let tempLabels = {};
+
+        if (issue.events) {
+            for (let event of issue.events) {
+                let eventDateTime: Date;
+                if (event.created_at) {
+                    eventDateTime = event.created_at;
+                }
+
+                if (event.event === 'labeled') {
+                    
+                }
+                //
+                // Process Project Stages
+                //
+                let toStage: string;
+                let toLevel: number;
+                let fromStage: string;
+                let fromLevel: number = 0;
+
+                if (event.project_card && event.project_card.column_name) {
+                    if (!addedTime) {
+                        addedTime = eventDateTime;
+                    }
+
+                    if (!event.project_card.stage_name) {
+                        throw new Error("stage_name should have been set already");
+                    }
+
+                    toStage = event.project_card.stage_name;
+                    toLevel = stageLevel[toStage];
+                    currentStage = toStage;
+                }
+        
+                if (event.project_card && event.project_card.previous_column_name) {
+                    if (!event.project_card.previous_stage_name) {
+                        throw new Error("previous_stage_name should have been set already");
+                    }
+
+                    fromStage = event.project_card.previous_stage_name;
+                    fromLevel = stageLevel[fromStage];
+                }
+
+                // last occurence of moving to these columns from a lesser or no column
+                // example. if moved to accepted from proposed (or less), 
+                //      then in-progress (greater) and then back to accepted, first wins            
+                if (toStage === 'Proposed' || toStage === 'Accepted' || toStage === 'In-Progress') {
+                    if (toLevel > fromLevel) {
+                        issue[this.stageAtNames[toLevel]] = eventDateTime;
+                        console.log(`${this.stageAtNames[toLevel]}: ${eventDateTime}`);
+                    } 
+                }
+
+                if (toStage === 'Done') {
+                    doneTime = eventDateTime;
+                }
+            }
+
+            // done_at and blocked_at is only set if it's currently at that stage
+            if (currentStage === 'Done') {
+                issue.project_done_at = doneTime;
+                console.log(`project_done_at: ${issue.project_done_at}`);
+            }
+
+            if (addedTime) {
+                issue.project_added_at = addedTime;
+                console.log(`project_added_at: ${issue.project_added_at}`);
+            }
+
+            issue.project_stage = currentStage;
+            console.log(`project_stage: ${issue.project_stage}`);
+        }
+    }   
 }
