+++ reports/project-done.ts
@@ -1,10 +1,13 @@
-import {ProjectIssue, CrawlingTarget} from '../interfaces';
-import {ProjectStages, ProjectStageIssues} from '../project-reports-lib';
+import {CrawlingTarget} from '../interfaces';
+import {ProjectIssue} from '../project-reports-lib';
 import * as rptLib from '../project-reports-lib';
 const tablemark = require('tablemark')
 import * as os from 'os';
+import moment = require('moment');
 
 let clone = require('clone');
+
+let now = moment();
 
 const reportType = 'project';
 export {reportType};
@@ -34,19 +37,18 @@ export function process(config: any, issues: ProjectIssue[], drillIn: (identifie
 
     completedCards.cardType = config["report-on"] || config["report-on-label"];
 
-    let daysAgo = config['daysAgo'];
-    if (isNaN(daysAgo)) {
+    let flagDaysAgo = config['daysAgo'] || 7;
+    if (isNaN(flagDaysAgo)) {
         throw new Error("daysAgo is not a number");
     }
-    completedCards.daysAgo = daysAgo;
+    completedCards.daysAgo = flagDaysAgo;
 
-    var dateAgo = new Date();
-    dateAgo.setDate(dateAgo.getDate() - config['daysAgo']);
+    let doneDaysAgo = now.diff(config['daysAgo'] || 7, 'days', true);
 
     console.log(`Getting cards for ${completedCards.cardType}`);
     let cardsForType = completedCards.cardType === '*'? clone(issues) : clone(rptLib.filterByLabel(issues, completedCards.cardType.toLowerCase()) as ProjectIssue[]);
 
-    completedCards.cards = cardsForType.filter(issue => issue["project_done_at"] && new Date(issue["project_done_at"]) > dateAgo);
+    completedCards.cards = cardsForType.filter(issue => issue["project_done_at"] && doneDaysAgo > flagDaysAgo);
 
     return completedCards;
 }
@@ -78,7 +80,7 @@ export function renderMarkdown(targets: CrawlingTarget[], processedData: any): s
 
         doneRow.assigned = assigned ? `<img height="20" width="20" alt="@${assigned.login}" src="${assigned.avatar_url}"/> <a href="${assigned.html_url}">${assigned.login}</a>` : ":triangular_flag_on_post:";
         doneRow.title = `[${card.title}](${card.html_url})`;
-        doneRow.completed = new Date(card["project_done_at"]).toDateString();
+        doneRow.completed = now.to(moment(card["project_done_at"]));
 
         rows.push(doneRow);
     }
